requirements.txt
playwright>=1.40
Pillow>=10.0
mss>=9.0
